requirements.txt
requests
python-dotenv
pyTelegramBotAPI
Django
